move supabase utils into shared dir
