+++ app/auth/actions.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { redirect } from "next/navigation";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@shared/utils/supabase/server";
 import { revalidatePath } from "next/cache";
 
 export async function signInWithGoogle() {
