+++ app/chatrooms/[chatroomId]/page.tsx
@@ -1,4 +1,4 @@
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@shared/utils/supabase/server";
 import Link from "next/link";
 import { redirect } from "next/navigation";
 import LeaveChatroomButton from "./components/leave-chatroom-button";
