+++ app/chatrooms/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@shared/utils/supabase/server";
 import { revalidatePath } from "next/cache";
 import {
   getRagflowDatasetId,
